Use Fig standards coding style, add composer.json and clean the repo
Use Fig standards (PSR-0, 1 and 2)
move php files to "lib/Testify"
add composer.json
add a simple "test.php"
add .gitignore
add alias "__invoke()" method with "run()"
