+++ Readme.md
@@ -6,8 +6,10 @@ Testify makes writing unit tests fun again. It has an elegant syntax and keeps t
 Here is an example for a test suite with two test cases:
 
 ```php
-include "testify/testify.class.php";
-include "MyCalc.php";
+require 'vendor/autoload.php';
+
+use Math\MyCalc;
+use Testify\Testify;
 
 $tf = new Testify("MyCalc Test Suite");
 
@@ -17,7 +19,7 @@ $tf->beforeEach(function($tf){
 
 $tf->test("Testing the add() method", function($tf){
 	$calc = $tf->data->calc;
-	
+
 	$calc->add(4);
 	$tf->assert($calc->result() == 14);
 
@@ -27,7 +29,7 @@ $tf->test("Testing the add() method", function($tf){
 
 $tf->test("Testing the mul() method", function($tf){
 	$calc = $tf->data->calc;
-	
+
 	$calc->mul(1.5);
 	$tf->assertEqual($calc->result(),15);
 
@@ -35,7 +37,7 @@ $tf->test("Testing the mul() method", function($tf){
 	$tf->assertEqual($calc->result(),-15);
 });
 
-$tf->run();
+$tf();
 ```
 
-For full documentation and a getting started guide, visit the [Testify homepage](http://tutorialzine.com/projects/testify/). 
+For full documentation and a getting started guide, visit the [Testify homepage](http://tutorialzine.com/projects/testify/).
