+++ composer.json
@@ -1,0 +1,15 @@
+{
+    "name": "bafs/testify",
+    "description": "Testify makes writing unit tests fun again. It has an elegant syntax and keeps things simple",
+    "type": "library",
+    "keywords": ["unit", "testing", "framework"],
+    "license": "GPL",
+    "require": {
+        "php": ">=5.3.0"
+    },
+    "autoload": {
+        "psr-0" : {
+            "Testify" : "lib/"
+        }
+    }
+}
