+++ examples/example2.php
@@ -2,27 +2,29 @@
 
 /*
  * A more complex example using beforeEach and data
- * 
+ *
  */
 
-require '../testify/testify.class.php';
+require '../vendor/autoload.php';
+
+use Testify\Testify;
 
 $tf = new Testify("A bit more advanced test suite");
 
 // Before each is called before every test case
 
-$tf->beforeEach(function($tf){
-	
+$tf->beforeEach(function($tf) {
+
 	// Use the data property to share variables across tests
-	
+
 	$tf->data->arr = array('a','b','c','d','e','f');
-	
+
 });
 
-$tf->test("Testing Array Pop", function($tf){
+$tf->test("Testing Array Pop", function($tf) {
 
 	$arr = &$tf->data->arr;
-	
+
 	$tf->assertEqual(array_pop($arr),'f');
 	$tf->assertEqual(array_pop($arr),'e');
 	$tf->assertEqual(array_pop($arr),'d');
@@ -30,12 +32,12 @@ $tf->test("Testing Array Pop", function($tf){
 
 });
 
-$tf->test("Testing In Array", function($tf){
+$tf->test("Testing In Array", function($tf) {
 
 	// beforeEach has restored the array
 
 	$arr = &$tf->data->arr;
-	
+
 	$tf->assertInArray('a',$arr);
 	$tf->assertInArray('b',$arr);
 	$tf->assertInArray('c',$arr);
@@ -49,5 +51,3 @@ $tf->test("Testing In Array", function($tf){
 include 'subtest.php';
 
 $tf->run();
-
-?>
