+++ examples/subtest.php
@@ -1,8 +1,6 @@
 <?php
 
-$tf->test("An independent test file",function($tf){
+$tf->test("An independent test file", function($tf) {
 	$tf->pass();
 	$tf->assert(true);
 });
-
-?>
